requirements.txt
astroid==2.2.5
bcrypt==3.1.6
cffi==1.12.3
Django==2.2.2
django-templated-mail==1.1.1
djangorestframework==3.9.4
djoser==1.7.0
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
psycopg2-binary==2.8.2
pycparser==2.19
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.1
